requirements.txt
mysql-connector-python==8.0.21
protobuf==3.12.2
six==1.15.0
urllib3==1.25.9
